refactor: LottoNumberGenerator를 LottoMachine에 주입
